feat : completion maintain history

+++ src/main/java/spring/library/service/BookService.java
@@ -5,7 +5,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import spring.library.controller.request.BookRequest;
 import spring.library.domain.Book;
+import spring.library.domain.BookHistory;
 import spring.library.dto.BookDto;
+import spring.library.repository.BookHistoryRepository;
 import spring.library.repository.BookRepository;
 
 import java.util.List;
@@ -14,11 +16,15 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BookService {
     private final BookRepository bookRepository;
+    private final BookHistoryRepository bookHistoryRepository;
 
     public BookDto save(BookRequest bookRequest) {
+        BookHistory bookHistory = BookHistory.from(bookRequest);
+        bookHistoryRepository.save(bookHistory);
         Book book = Book.from(bookRequest);
-        BookDto bookDto = BookDto.from(bookRepository.save(book));
-        return bookDto;
+        bookRepository.save(book);
+
+        return BookDto.from(book);
     }
 
     public List<BookDto> load(){
@@ -29,13 +35,24 @@ public class BookService {
 
     @Transactional
     public BookDto update(Long bookId, BookRequest BookRequest) {
-        Book bookOfTitle = bookRepository.findById(bookId).orElseThrow(() -> new IllegalArgumentException("null"));
+        Book book = bookRepository.findById(bookId).orElseThrow(() -> new IllegalArgumentException("There is not book for endtered bookId"));
+        BookHistory bookHistory = bookHistoryRepository.findById(bookId).orElseThrow(() -> new IllegalArgumentException("null"));
+
         BookDto bookDto = BookDto.from(BookRequest);
-        bookOfTitle.update(bookDto);
+
+        bookHistory.update(bookDto);
+        book.update(bookDto);
+
+        bookHistoryRepository.save(bookHistory);
+        bookRepository.save(book);
         return bookDto;
     }
 
     public void delete(Long bookId) {
+        Book book = bookRepository.findById(bookId).orElseThrow(() -> new IllegalArgumentException("There is not book for endtered bookId"));
+        if(book.getStatus().equals("대출중")){
+            throw new IllegalArgumentException("you cannot delete book : this book is loaned");
+        }
         bookRepository.deleteById(bookId);
     }
 }
